Revert "fix compilation errors"
This reverts commit 0deb09b1b4588ef32112ff68c05e5c6f5761af67.

We need to keep sdno-testframework under the test/ directory
for now since "Test" is one of the sub-projects in the
Integration project proposal, and there will also be other
test frameworks in use.  We can reorganize the directory
structures once we have a better understanding of how different
test frameworks will be used by various projects.

Change-Id: I42597aaed369c578a65386c8c8e563e9d44e310d
Signed-off-by: Gary Wu <gary.i.wu@huawei.com>