Fix scripts bugs
Issue-id: SDNO-1475
Problems related to start and pull scripts: trying to delete docker
instance if there is any running without checking if they exist in
reality, there is no impact of the execution but it is a misleading
information once somebody is trying to debug any issue.
the satart image now include a wait for lcm/nslcm docker to be up and
runnig so the setup and tear down of tests (sdno-lcm) are now adapted
to that.
PS: the sdno-nslcm/sanity-check-setup.sh has been changed by mistake.
roll back to the original version.

Change-Id: I2eb34c3c2bf663e09ed37c94556e198058d9ec88
Signed-off-by: Karim Nabli <karim.nabli@huawei.com>